clean up packages
