+++ annotation-processors/src/main/kotlin/at/hannibal2/skyhanni/kmixin/injectors/inject/InjectAtSerializer.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.kmixin.injectors
+package at.hannibal2.skyhanni.kmixin.injectors.inject
 
 import at.hannibal2.skyhanni.kmixin.addModifiers
 import at.hannibal2.skyhanni.kmixin.addParameter
@@ -8,9 +8,12 @@ import at.hannibal2.skyhanni.kmixin.annotations.KSelf
 import at.hannibal2.skyhanni.kmixin.annotations.KShadow
 import at.hannibal2.skyhanni.kmixin.annotations.KStatic
 import at.hannibal2.skyhanni.kmixin.annotations.getAsBoolean
-import at.hannibal2.skyhanni.kmixin.annotations.getAsInjectionKind
+import at.hannibal2.skyhanni.kmixin.annotations.getAsInt
 import at.hannibal2.skyhanni.kmixin.annotations.getAsString
+import at.hannibal2.skyhanni.kmixin.annotations.getAsTargetLocation
 import at.hannibal2.skyhanni.kmixin.hasAnnotation
+import at.hannibal2.skyhanni.kmixin.injectors.InjectionSerializer
+import at.hannibal2.skyhanni.kmixin.injectors.InjectionUtils
 import at.hannibal2.skyhanni.kmixin.toJava
 import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
@@ -20,14 +23,17 @@ import com.squareup.javapoet.FieldSpec
 import com.squareup.javapoet.MethodSpec
 import javax.lang.model.element.Modifier
 
-object InjectSerializer : InjectionSerializer {
+object InjectAtSerializer : InjectionSerializer {
 
     private const val CAPTURE_FAILHARD = "org.spongepowered.asm.mixin.injection.callback.LocalCapture.CAPTURE_FAILHARD"
 
     override fun readAnnotation(function: KSFunctionDeclaration, annotation: KSAnnotation): AnnotationSpec = with(annotation) {
+        val shift = "org.spongepowered.asm.mixin.injection.At.Shift.${getAsTargetLocation("shift").name}"
+        val atString = "@\$T(value = \"INVOKE\", ordinal = \$L, shift = ${shift}, target = \$S)"
+
         AnnotationSpec.builder(INJECT_CLASS)
             .addMember("method", "\"${getAsString("method")}\"")
-            .addMember("at", "@\$T(value = \"${getAsInjectionKind("kind").name}\")", AT_CLASS)
+            .addMember("at", atString, AT_CLASS, getAsInt("ordinal"), getAsString("target"))
             .addMember("cancellable", "\$L", getAsBoolean("cancellable"))
             .addMember("remap", "\$L", getAsBoolean("remap"))
             .apply {
