+++ annotation-processors/src/main/kotlin/at/hannibal2/skyhanni/kmixin/injectors/inject/InjectSerializer.kt
@@ -1,16 +1,18 @@
-package at.hannibal2.skyhanni.kmixin.injectors
+package at.hannibal2.skyhanni.kmixin.injectors.inject
 
-import at.hannibal2.skyhanni.kmixin.addAnnotation
 import at.hannibal2.skyhanni.kmixin.addModifiers
 import at.hannibal2.skyhanni.kmixin.addParameter
 import at.hannibal2.skyhanni.kmixin.annotations.AT_CLASS
+import at.hannibal2.skyhanni.kmixin.annotations.INJECT_CLASS
+import at.hannibal2.skyhanni.kmixin.annotations.KSelf
 import at.hannibal2.skyhanni.kmixin.annotations.KShadow
 import at.hannibal2.skyhanni.kmixin.annotations.KStatic
-import at.hannibal2.skyhanni.kmixin.annotations.REDIRECT_CLASS
 import at.hannibal2.skyhanni.kmixin.annotations.getAsBoolean
-import at.hannibal2.skyhanni.kmixin.annotations.getAsInt
+import at.hannibal2.skyhanni.kmixin.annotations.getAsInjectionKind
 import at.hannibal2.skyhanni.kmixin.annotations.getAsString
 import at.hannibal2.skyhanni.kmixin.hasAnnotation
+import at.hannibal2.skyhanni.kmixin.injectors.InjectionSerializer
+import at.hannibal2.skyhanni.kmixin.injectors.InjectionUtils
 import at.hannibal2.skyhanni.kmixin.toJava
 import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
@@ -18,20 +20,21 @@ import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.squareup.javapoet.AnnotationSpec
 import com.squareup.javapoet.FieldSpec
 import com.squareup.javapoet.MethodSpec
-import com.squareup.javapoet.TypeName
 import javax.lang.model.element.Modifier
 
-object RedirectMethodSerializer : InjectionSerializer {
+object InjectSerializer : InjectionSerializer {
+
+    private const val CAPTURE_FAILHARD = "org.spongepowered.asm.mixin.injection.callback.LocalCapture.CAPTURE_FAILHARD"
 
     override fun readAnnotation(function: KSFunctionDeclaration, annotation: KSAnnotation): AnnotationSpec = with(annotation) {
-        AnnotationSpec.builder(REDIRECT_CLASS)
-            .addMember("method", "\$S", getAsString("method"))
-            .addAnnotation("at", AT_CLASS) {
-                add("value", "\$S", "INVOKE")
-                add("target", "\$S", getAsString("target"))
-                add("ordinal", "\$L", getAsInt("ordinal"))
-            }
+        AnnotationSpec.builder(INJECT_CLASS)
+            .addMember("method", "\"${getAsString("method")}\"")
+            .addMember("at", "@\$T(value = \"${getAsInjectionKind("kind").name}\")", AT_CLASS)
+            .addMember("cancellable", "\$L", getAsBoolean("cancellable"))
             .addMember("remap", "\$L", getAsBoolean("remap"))
+            .apply {
+                if (getAsBoolean("captureLocals")) addMember("locals", CAPTURE_FAILHARD)
+            }
             .build()
     }
 
@@ -40,30 +43,27 @@ object RedirectMethodSerializer : InjectionSerializer {
         annotation: AnnotationSpec,
         function: KSFunctionDeclaration,
         methodWriter: (MethodSpec.Builder) -> Unit,
-        fieldWriter: (FieldSpec.Builder) -> Unit
+        fieldWriter: (FieldSpec.Builder) -> Unit,
     ) {
-        val returnType = function.returnType!!.toJava()
         val method = MethodSpec.methodBuilder(function.simpleName.asString())
             .addModifiers(Modifier.PRIVATE)
             .addModifiers(function.hasAnnotation(KStatic::class), Modifier.STATIC)
             .addAnnotation(annotation)
-            .returns(returnType)
+            .returns(Void.TYPE)
 
         function.parameters
-            .filter { !it.hasAnnotation(KShadow::class) }
+            .filter { !it.hasAnnotation(KShadow::class) && !it.hasAnnotation(KSelf::class) }
             .forEach {
                 require(!it.hasDefault) { "Default parameters are not supported" }
                 method.addParameter(it)
             }
 
-        if (returnType == TypeName.VOID) {
-            method.addStatement("\$T.INSTANCE.${function.simpleName.asString()}(${InjectionUtils.createParameterList(function)})", klass.toJava())
-        } else {
-            method.addStatement("return \$T.INSTANCE.${function.simpleName.asString()}(${InjectionUtils.createParameterList(function)})", klass.toJava())
-        }
+        method.addStatement(
+            "\$T.INSTANCE.${function.simpleName.asString()}(${InjectionUtils.createParameterList(function)})",
+            klass.toJava()
+        )
 
         methodWriter(method)
         InjectionUtils.gatherShadows(function, fieldWriter, methodWriter)
     }
-
 }
