+++ annotation-processors/src/main/kotlin/at/hannibal2/skyhanni/kmixin/injectors/redirect/RedirectMethodSerializer.kt
@@ -1,17 +1,18 @@
-package at.hannibal2.skyhanni.kmixin.injectors
+package at.hannibal2.skyhanni.kmixin.injectors.redirect
 
+import at.hannibal2.skyhanni.kmixin.addAnnotation
 import at.hannibal2.skyhanni.kmixin.addModifiers
 import at.hannibal2.skyhanni.kmixin.addParameter
 import at.hannibal2.skyhanni.kmixin.annotations.AT_CLASS
-import at.hannibal2.skyhanni.kmixin.annotations.INJECT_CLASS
-import at.hannibal2.skyhanni.kmixin.annotations.KSelf
 import at.hannibal2.skyhanni.kmixin.annotations.KShadow
 import at.hannibal2.skyhanni.kmixin.annotations.KStatic
+import at.hannibal2.skyhanni.kmixin.annotations.REDIRECT_CLASS
 import at.hannibal2.skyhanni.kmixin.annotations.getAsBoolean
 import at.hannibal2.skyhanni.kmixin.annotations.getAsInt
 import at.hannibal2.skyhanni.kmixin.annotations.getAsString
-import at.hannibal2.skyhanni.kmixin.annotations.getAsTargetLocation
 import at.hannibal2.skyhanni.kmixin.hasAnnotation
+import at.hannibal2.skyhanni.kmixin.injectors.InjectionSerializer
+import at.hannibal2.skyhanni.kmixin.injectors.InjectionUtils
 import at.hannibal2.skyhanni.kmixin.toJava
 import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
@@ -19,24 +20,20 @@ import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.squareup.javapoet.AnnotationSpec
 import com.squareup.javapoet.FieldSpec
 import com.squareup.javapoet.MethodSpec
+import com.squareup.javapoet.TypeName
 import javax.lang.model.element.Modifier
 
-object InjectAtSerializer : InjectionSerializer {
-
-    private const val CAPTURE_FAILHARD = "org.spongepowered.asm.mixin.injection.callback.LocalCapture.CAPTURE_FAILHARD"
+object RedirectMethodSerializer : InjectionSerializer {
 
     override fun readAnnotation(function: KSFunctionDeclaration, annotation: KSAnnotation): AnnotationSpec = with(annotation) {
-        val shift = "org.spongepowered.asm.mixin.injection.At.Shift.${getAsTargetLocation("shift").name}"
-        val atString = "@\$T(value = \"INVOKE\", ordinal = \$L, shift = ${shift}, target = \$S)"
-
-        AnnotationSpec.builder(INJECT_CLASS)
-            .addMember("method", "\"${getAsString("method")}\"")
-            .addMember("at", atString, AT_CLASS, getAsInt("ordinal"), getAsString("target"))
-            .addMember("cancellable", "\$L", getAsBoolean("cancellable"))
-            .addMember("remap", "\$L", getAsBoolean("remap"))
-            .apply {
-                if (getAsBoolean("captureLocals")) addMember("locals", CAPTURE_FAILHARD)
+        AnnotationSpec.builder(REDIRECT_CLASS)
+            .addMember("method", "\$S", getAsString("method"))
+            .addAnnotation("at", AT_CLASS) {
+                add("value", "\$S", "INVOKE")
+                add("target", "\$S", getAsString("target"))
+                add("ordinal", "\$L", getAsInt("ordinal"))
             }
+            .addMember("remap", "\$L", getAsBoolean("remap"))
             .build()
     }
 
@@ -45,27 +42,38 @@ object InjectAtSerializer : InjectionSerializer {
         annotation: AnnotationSpec,
         function: KSFunctionDeclaration,
         methodWriter: (MethodSpec.Builder) -> Unit,
-        fieldWriter: (FieldSpec.Builder) -> Unit,
+        fieldWriter: (FieldSpec.Builder) -> Unit
     ) {
+        val returnType = function.returnType!!.toJava()
         val method = MethodSpec.methodBuilder(function.simpleName.asString())
             .addModifiers(Modifier.PRIVATE)
             .addModifiers(function.hasAnnotation(KStatic::class), Modifier.STATIC)
             .addAnnotation(annotation)
-            .returns(Void.TYPE)
+            .returns(returnType)
 
         function.parameters
-            .filter { !it.hasAnnotation(KShadow::class) && !it.hasAnnotation(KSelf::class) }
+            .filter { !it.hasAnnotation(KShadow::class) }
             .forEach {
                 require(!it.hasDefault) { "Default parameters are not supported" }
                 method.addParameter(it)
             }
 
-        method.addStatement(
-            "\$T.INSTANCE.${function.simpleName.asString()}(${InjectionUtils.createParameterList(function)})",
-            klass.toJava()
-        )
+        if (returnType == TypeName.VOID) {
+            method.addStatement("\$T.INSTANCE.${function.simpleName.asString()}(${
+                InjectionUtils.createParameterList(
+                    function
+                )
+            })", klass.toJava())
+        } else {
+            method.addStatement("return \$T.INSTANCE.${function.simpleName.asString()}(${
+                InjectionUtils.createParameterList(
+                    function
+                )
+            })", klass.toJava())
+        }
 
         methodWriter(method)
         InjectionUtils.gatherShadows(function, fieldWriter, methodWriter)
     }
+
 }
